+++ bio_embeddings/embed/prottrans_embedder.py
@@ -1,3 +1,4 @@
+import abc
 import logging
 import re
 from itertools import zip_longest
@@ -12,15 +13,16 @@ from bio_embeddings.embed.embedder_interfaces import EmbedderWithFallback
 logger = logging.getLogger(__name__)
 
 
-class ProtTransT5BFDEmbedder(EmbedderWithFallback):
-    """Encoder of the ProtTrans T5 BFD model
+class ProtTransT5Embedder(EmbedderWithFallback, abc.ABC):
+    """Encoder of the ProtTrans T5 model, both BFD and BFD finetuned on UniRef50. To embed please pick either
+    ProtTransT5BFDEmbedder or ProtTransT5UniRef50Embedder
 
     Note that this model alone takes 13GB, so you need a GPU with a lot of memory.
     """
 
     _model: T5EncoderModel
     _decoder: bool = False
-    name = "prottrans_t5_bfd"
+    _half_model: bool = False
     embedding_dimension = 1024
     number_of_layers = 1
     necessary_directories = ["model_directory"]
@@ -38,12 +40,16 @@ class ProtTransT5BFDEmbedder(EmbedderWithFallback):
         # Should the need arise we can just split this class in to an encoder and a decoder subclass
         # by setting one subclass to _decoder=True and the other to _decoder=False
         self._decoder = self._options.get("decoder", False)
+        self._half_model = self._options.get("half_model", False)
 
         # make model
         if self._decoder:
             self._model = T5EncoderModel.from_pretrained(self._model_directory)
         else:
             self._model = T5Model.from_pretrained(self._model_directory)
+        # Compute in half precision, saving us half the memory
+        if self._half_model:
+            self._model.half()
         self._model = self._model.to(self._device).eval()
         self._model_fallback = None
         self._tokenizer = T5Tokenizer.from_pretrained(
@@ -114,3 +120,25 @@ class ProtTransT5BFDEmbedder(EmbedderWithFallback):
     def embed(self, sequence: str) -> ndarray:
         [embedding] = self.embed_batch([sequence])
         return embedding
+
+
+class ProtTransT5BFDEmbedder(ProtTransT5Embedder):
+    """Encoder of the ProtTrans T5 model trained on BFD
+
+    We recommend settings `half_model=True`, which on the tested GPU (Quadro RTX 3000) reduces memory consumption
+    from 12GB to 7GB while the effect in benchmark is negligible (±0.1 percentages points in different sets,
+    generally below standard error)
+    """
+
+    name = "prottrans_t5_bfd"
+
+
+class ProtTransT5UniRef50Embedder(ProtTransT5Embedder):
+    """Encoder of the ProtTrans T5 model trained on BFD and finetuned on UniRef 50
+
+    We recommend settings `half_model=True`, which on the tested GPU (Quadro RTX 3000) reduces memory consumption
+    from 12GB to 7GB while the effect in benchmark is negligible (±0.1 percentages points in different sets,
+    generally below standard error)
+    """
+
+    name = "prottrans_t5_uniref50"