+++ tests/test_embedder.py
@@ -22,6 +22,7 @@ from unittest import mock
 import numpy
 import pytest
 import torch
+from numpy import ndarray
 
 from bio_embeddings.embed import (
     BeplerEmbedder,
@@ -33,11 +34,13 @@ from bio_embeddings.embed import (
     ProtTransAlbertBFDEmbedder,
     ProtTransBertBFDEmbedder,
     ProtTransT5BFDEmbedder,
+    ProtTransT5UniRef50Embedder,
     ProtTransXLNetUniRef100Embedder,
     SeqVecEmbedder,
     UniRepEmbedder,
 )
-from bio_embeddings.utilities import read_fasta
+from bio_embeddings.embed.pipeline import embed_and_write_batched
+from bio_embeddings.utilities import read_fasta, FileSystemFileManager
 from tests.shared import check_embedding
 
 all_embedders = [
@@ -50,6 +53,12 @@ all_embedders = [
     ProtTransBertBFDEmbedder,
     pytest.param(
         ProtTransT5BFDEmbedder,
+        marks=pytest.mark.skipif(
+            os.environ.get("SKIP_T5"), reason="T5 makes ci run out of disk"
+        ),
+    ),
+    pytest.param(
+        ProtTransT5UniRef50Embedder,
         marks=pytest.mark.skipif(
             os.environ.get("SKIP_T5"), reason="T5 makes ci run out of disk"
         ),
@@ -207,3 +216,41 @@ def test_batching_t5(pytestconfig):
         assert not numpy.allclose(a, b) and numpy.allclose(
             a, b, rtol=1.0e-4, atol=1.0e-5
         )
+
+
+def test_half_precision_embedder(pytestconfig, caplog, tmp_path: Path):
+    """Currently a dummy test"""
+    class Float16Embedder(EmbedderInterface):
+        name = "float16embedder"
+        embedding_dimension = 1024
+        number_of_layers = 1
+
+        def __init__(self, **kwargs):
+            super().__init__(**kwargs)
+            assert kwargs.get("half_model"), kwargs
+
+        def embed(self, sequence: str) -> ndarray:
+            return numpy.random.random((len(sequence), 1024)).astype(numpy.float16)
+
+        @staticmethod
+        def reduce_per_protein(embedding: ndarray) -> ndarray:
+            return embedding.sum(axis=0)
+
+    result_kwargs = {
+        "prefix": str(tmp_path),
+        "remapped_sequences_file": str(
+            pytestconfig.rootpath.joinpath("test-data/remapped_sequences_file.fasta")
+        ),
+        "mapping_file": str(
+            pytestconfig.rootpath.joinpath("test-data/mapping_file.csv")
+        ),
+        "half_model": True,
+    }
+    embed_and_write_batched(
+        Float16Embedder(**result_kwargs),
+        FileSystemFileManager(),
+        result_kwargs=result_kwargs,
+    )
+
+    assert caplog.messages == []
+